Initial split of tokenization transform into ray and python
