+++ transforms/universal/filter/ray/README.md
@@ -1,4 +1,4 @@
-# Filtering
+# Filtering Transform for Ray
 Please see the set of
 [transform project conventions](../../../README.md)
 for details on general project conventions, transform configuration,
