+++ transforms/universal/tokenization/kfp_ray/v1/tokenization_wf.py
@@ -23,10 +23,10 @@ from kfp_support.workflow_support.utils import (
 # the name of the job script
 EXEC_SCRIPT_NAME: str = "tokenization_transform_ray.py"
 
-task_image = "quay.io/dataprep1/data-prep-kit/tokenization-ray:0.3.0"
+task_image = "quay.io/dataprep1/data-prep-kit/tokenization-ray:0.3.0.dev6"
 
 # components
-base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.2.0"
+base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:0.2.0.dev6"
 # path to kfp component specifications files
 
 component_spec_path = "../../../../../kfp/kfp_ray_components/"
