+++ transforms/universal/tokenization/python/.dockerignore
@@ -1,0 +1,1 @@
+venv/
