+++ transforms/universal/tokenization/python/test-data/ds01/expected/metadata.json
@@ -1,0 +1,59 @@
+{
+    "pipeline": "pipeline_id",
+    "job details": {
+        "job category": "preprocessing",
+        "job name": "Tokenization",
+        "job type": "ray",
+        "job id": "job_id",
+        "start_time": "2024-03-29 13:30:56",
+        "end_time": "2024-03-29 13:30:57",
+        "status": "success"
+    },
+    "code": {
+        "github": "github",
+        "commit_hash": "12345",
+        "path": "path"
+    },
+    "job_input_params": {
+        "tokenizer": "hf-internal-testing/llama-tokenizer",
+        "tokenizer_args": null,
+        "doc_id_column": "document_id",
+        "doc_content_column": "contents",
+        "text_lang": "en",
+        "chunk_size": 0,
+        "checkpointing": false,
+        "max_files": -1,
+        "number of workers": 5,
+        "worker options": {
+            "num_cpus": 0.8
+        },
+        "actor creation delay": 0
+    },
+    "execution_stats": {
+        "cpus": 10,
+        "gpus": 0,
+        "memory": 27.31659088190645,
+        "object_store": 2.0
+    },
+    "job_output_stats": {
+        "source_files": 5,
+        "source_size": 450,
+        "result_files": 3,
+        "result_size": 842,
+        "table_processing": 0.03880786895751953,
+        "num_files": 3,
+        "num_rows": 6,
+        "num_tokenized_rows": 6,
+        "num_tokens": 85,
+        "num_chars": 384,
+        "skipped empty tables": 2
+    },
+    "source": {
+        "name": "/Users/xdang/00proj/04-FM/01_code/fm-data-engineering/transforms/universal/tokenization/test-data/ds01/input",
+        "type": "path"
+    },
+    "target": {
+        "name": "/Users/xdang/00proj/04-FM/01_code/fm-data-engineering/transforms/universal/tokenization/output/ds01",
+        "type": "path"
+    }
+}
