+++ transforms/universal/tokenization/python/test-data/ds02/expected/metadata.json
@@ -1,0 +1,58 @@
+{
+    "pipeline": "pipeline_id",
+    "job details": {
+        "job category": "preprocessing",
+        "job name": "Tokenization",
+        "job type": "ray",
+        "job id": "job_id",
+        "start_time": "2024-03-29 14:03:15",
+        "end_time": "2024-03-29 14:03:32",
+        "status": "success"
+    },
+    "code": {
+        "github": "github",
+        "commit_hash": "12345",
+        "path": "path"
+    },
+    "job_input_params": {
+        "tokenizer": "hf-internal-testing/llama-tokenizer",
+        "tokenizer_args": null,
+        "doc_id_column": "document_id",
+        "doc_content_column": "contents",
+        "text_lang": "en",
+        "chunk_size": 20000,
+        "checkpointing": false,
+        "max_files": -1,
+        "number of workers": 5,
+        "worker options": {
+            "num_cpus": 0.8
+        },
+        "actor creation delay": 0
+    },
+    "execution_stats": {
+        "cpus": 10,
+        "gpus": 0,
+        "memory": 27.180484008975327,
+        "object_store": 2.0
+    },
+    "job_output_stats": {
+        "source_files": 1,
+        "source_size": 16863266,
+        "result_files": 1,
+        "result_size": 37109764,
+        "table_processing": 15.886597871780396,
+        "num_files": 1,
+        "num_rows": 1,
+        "num_tokenized_rows": 1,
+        "num_tokens": 4638717,
+        "num_chars": 16836009
+    },
+    "source": {
+        "name": "/Users/xdang/00proj/04-FM/01_code/fm-data-engineering/transforms/universal/tokenization/test-data/ds02/input",
+        "type": "path"
+    },
+    "target": {
+        "name": "/Users/xdang/00proj/04-FM/01_code/fm-data-engineering/transforms/universal/tokenization/output/ds02",
+        "type": "path"
+    }
+}
