--- transforms/universal/tokenization/ray/requirements.txt
@@ -1,4 +1,0 @@
-# transform runtime
-#data-prep-kit==0.1.2
-# for downloading + loading tokenizers:
-transformers==4.38.0
